requirements.txt
streamlit
numpy
pandas
scikit-learn
scikit-surprise
pyngrok
